refactor : validation 로직 분리 및 패키지 정리
- validation 로직을 모듈화
  - 모듈화를 위해 각 검증 로직에 요구하는 매개변수 추가
  - signin.js 에서 validation 모듈을 불러와서 사용
- 패키지 정리
  - root 에 html 파일들 모아놓음
  - /script 에 js 파일 보관
  - /src 에 css 파일 보관
