+++ index.html
@@ -22,7 +22,7 @@
         <a href="index.html">
           <img src="images/logo.svg" alt="홈 버튼" />
         </a>
-        <a class="cta cta-login" href="/signin/signin.html">
+        <a class="cta cta-login" href="/signin.html">
           <span>로그인</span>
         </a>
       </div>
@@ -34,7 +34,7 @@
             <span class="highlight">세상의 모든 정보</span>를<br />
             쉽게 저장하고 관리해 보세요.
           </h1>
-          <a class="cta cta-add" href="signup/signup.html"> 링크 추가하기 </a>
+          <a class="cta cta-add" href="/signup.html"> 링크 추가하기 </a>
           <img class="hero-image" src="images/hero.png" alt="메인 이미지" />
         </section>
         <section class="contents store">
