+++ script/signin.js
@@ -1,3 +1,9 @@
+import {
+  validateInput,
+  validateEmail,
+  validatePassword,
+} from "./validation.js";
+
 const signinBox = document.querySelector(".signin-box");
 const signinBtn = document.querySelector(".signin-btn");
 
@@ -7,46 +13,11 @@ const passwordInput = document.querySelector(".sign-input-password");
 const emailErrorMsg = document.querySelector(".email-error-msg");
 const passwordErrorMsg = document.querySelector(".password-error-msg");
 
-const TEST_EMAIL = "test@codeit.com";
-const TEST_PW = "codeit101";
-
-const EMPTY_EMAIL_MSG = "이메일을 입력해주세요.";
-const INVALID_EMAIL_MSG = "올바른 이메일 주소가 아닙니다.";
 const LOGIN_FAIL_MESSAGE_EMAIL = "이메일을 확인해주세요.";
 
-const EMPTY_PASSWORD_MSG = "비밀번호를 입력해주세요.";
 const LOGIN_FAIL_MESSAGE_PASSWORD = "비밀번호를 확인해주세요.";
 
 const URL = "https://bootcamp-api.codeit.kr/api/sign-in";
-
-function validateInput(inputEl, errorMsgEl, message) {
-  errorMsgEl.textContent = message;
-  if (message) {
-    inputEl.classList.add("error");
-    return;
-  }
-  inputEl.classList.remove("error");
-  return;
-}
-
-function validateEmail() {
-  const email_regex = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,4}$/i;
-
-  if (emailInput.value === "") {
-    return validateInput(emailInput, emailErrorMsg, EMPTY_EMAIL_MSG);
-  }
-  if (!email_regex.test(emailInput.value)) {
-    return validateInput(emailInput, emailErrorMsg, INVALID_EMAIL_MSG);
-  }
-  return validateInput(emailInput, emailErrorMsg, "");
-}
-
-function validatePassword() {
-  if (passwordInput.value === "") {
-    return validateInput(passwordInput, passwordErrorMsg, EMPTY_PASSWORD_MSG);
-  }
-  return validateInput(passwordInput, passwordErrorMsg, "");
-}
 
 async function signin() {
   const request = {
@@ -78,14 +49,18 @@ async function signin() {
   }
 }
 
-function validationByEnter(e) {
+function signinByEnter(e) {
   if (e.key === "Enter") {
     e.preventDefault();
     signin();
   }
 }
 
-emailInput.addEventListener("focusout", validateEmail);
-passwordInput.addEventListener("focusout", validatePassword);
+emailInput.addEventListener("focusout", () =>
+  validateEmail(emailInput, emailErrorMsg)
+);
+passwordInput.addEventListener("focusout", () =>
+  validatePassword(passwordInput, passwordErrorMsg)
+);
 signinBtn.addEventListener("click", signin);
-signinBox.addEventListener("keydown", validationByEnter);
+signinBox.addEventListener("keydown", signinByEnter);
