+++ src/signin.css
@@ -1,0 +1,105 @@
+* {
+  box-sizing: border-box;
+}
+
+body {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 100px;
+  background-color: #f0f6ff;
+}
+
+input {
+  width: 100%;
+  padding: 18px 15px;
+  color: var(--black, #000);
+  font-family: Pretendard;
+  font-size: 1rem;
+  font-style: normal;
+  font-weight: 400;
+  line-height: normal;
+  border-radius: 8px;
+  border: 0.1px solid var(--linkbrary-gray-20, #ccd5e3);
+  background: var(--linkbrary-white, #fff);
+}
+
+input:focus {
+  border: 1px solid var(--linkbrary-primary-color, #6d6afe);
+  outline: none;
+}
+
+input.error {
+  border: 1px solid var(--linkbrary-primary-color, red);
+}
+
+span {
+  color: red;
+}
+
+.signin-box {
+  width: 30rem;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  gap: 30px;
+}
+
+.header {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 16px;
+}
+
+.sign-input-label {
+  font-size: 1rem;
+  font-weight: 300;
+}
+
+.sign-password {
+  position: relative;
+}
+
+.eye-button {
+  position: absolute;
+  bottom: 0.9rem;
+  right: 1rem;
+  width: 1.6rem;
+  height: 1.6rem;
+  border: none;
+  background-color: #fff;
+}
+
+.signin-btn {
+  width: 100%;
+  color: #f5f5f5;
+  display: flex;
+  padding: 16px 20px;
+  text-decoration: none;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+  gap: 10px;
+  border: none;
+  border-radius: 8px;
+  background: var(
+    --gra-purpleblue-to-skyblue,
+    linear-gradient(91deg, #6d6afe 0.12%, #6ae3fe 101.84%)
+  );
+}
+
+.social-login-box {
+  padding: 12px 24px;
+  border-radius: 8px;
+  border: 1px solid var(--linkbrary-gray-20, #ccd5e3);
+  background: var(--linkbrary-gray-10, #e7effb);
+}
+
+.social-login-inner {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
